moved isnput to input base class
